fix(wezterm): improper terminfo file setup
My local neovim setup has had scrolling bugs for like a month now, and
it has been just tolerable enough to not do anything about it. I finally
hit the breaking point and decided to look into it...

Turns out I have had this set up all wrong (or maybe it was right and
then it was changed?) for Nix. This should be more self-contained to
this setup.