feat(_common_): install devenv
devenv has some overlap with flakes, but I think devenv is more useful
overall many circumstances:

- terse configuration
- dev shells
- specific language support
- `npm run` analogue
- & more

It CANNOT build custom packages or secure docker containers, but if we
need that then a flake can be written for those purposes. `devenv` has
integration with flakes for this purpose, but we could also just ditch
the devenv at that point.

For docker containers, melange+apko seem like a better solution. I'd
say, in order of preference:

- melange+apko
- Nix
- dockerfiles

Either way, it seems like devenv is a better starting point for
most projects.