ref(neovim)!: port to 'main' nvim-treesitter
nvim-treesitter had a bunch of breaking changes. They removed a lot of
the convenience setup functions, and outsourced those to other plugins.

But I'm not going to install a bunch of plugins to manage my plugins.
I'm just going to write the code to manage it myself.

This changes a lot, because there were a lot of files with TS code that
must be isolated now. While doing so, it seemed prudent to do the same
for LSP.