feat(neovim): lightweight dressing.nvim replacement
I saw that dressing was archived. I wanted to swap to snacks, but that
is pretty unmaintained lately. Even when it is back to maintenance, I'm
not sure how much I'll trust it long-term.

I looked at the code for dressing and it was only a few hundred lines. A
lot of it was extra features and use cases I have no intention of
supporting/need. I was able to re-write just the ui.input portion
pretty easily in an afternoon, and am thus moving off of dressing.