ref!: move integration locations
Before, integrations would be located in the "parent" program
(e.g. an integration WITH fish FOR ripgrep is stored with the fish
configs) and would be conditionally located if the "subject" program
was installed.

This led to a lot, LOT of recursive bugs.

Now, we just unconditionally set up integrations if the "subject"
program is installed. Who cares if some fish functions get setup if fish
never gets installed? they'll never end up in the user env because nix
is smart like that.