ref(sway): move swayr->fuzzel integration to fuzzel->swayr integration
I think of "integrations" as a piece of configuration
you can provide to another thing, and if that other thing is not
installed, then there is no problem.

The way we did this before would have been broken if swayr was enabled
but not fuzzel. Now, swayr is safe to enable without fuzzel, and vice
versa.