rev(common)!: uninstall ast-grep
Man, this thing sounded so cool on paper.
But I really don't know when I would ever actually use this, as it is.
The premise is so cool though :(

I just could never justify fiddling with yaml to grep or refactor when I
could just... grep or refactor something.

If the CLI was more featured, it would be probably one of the best tools
in my toolbox.